==> model/03-language-modelling.ipynb <==
# Import Library

import joblib
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity

from sklearn.model_selection import train_test_split
from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import RandomizedSearchCV, GridSearchCV
from sklearn.model_selection import learning_curve

from sklearn.linear_model import LogisticRegression
from sklearn.multioutput import MultiOutputClassifier

from sklearn.metrics import classification_report
from sklearn.metrics import accuracy_score

import warnings
warnings.filterwarnings('ignore')

# Load Dataset
path = '../dataset/data_modelling/data_preprocessing.csv'
dataset = pd.read_csv(path, index_col=None)
dataset.head()

dataset.info()

dataset = dataset.drop(['Tweet', 'Casefolding', 'Char_Length_Prep'], axis=1)
dataset = dataset[['Stopwords', 'HS', 'Abusive']]
dataset = dataset.rename(columns={'Stopwords':'Tweet'})
dataset.head()

# Dataset Splitting

# Assignment Variabel X, y

X = dataset.Tweet
y = dataset.drop(columns='Tweet')
X.shape, y.shape

# train_test_split

X_train, X_test, y_train, y_test = train_test_split(X,
                                                    y,
                                                    stratify=y,
                                                    shuffle=True,
                                                    test_size=0.3,
                                                    random_state=42)

X_train.shape, X_test.shape, y_train.shape, y_test.shape

# TFIDF Vectorizer

from luwiji.text_proc import illustration, demo
demo.text

illustration.structured

illustration.bag_of_words

illustration.inverse_df

illustration.practical_idf

# Build Vector

tfidf = TfidfVectorizer(max_features=5000, ngram_range=(1,3))
X_train_tfidf = tfidf.fit(X_train.values.astype('U'))

X_train_tfidf = tfidf.transform(X_train.values.astype('U'))
X_test_tfidf = tfidf.transform(X_test.values.astype('U'))

df = pd.DataFrame(X_train_tfidf.todense().T,
                 index = tfidf.get_feature_names(),
                 columns=[f'D{i+1}' for i in range(len(X_train))])
df

# assignment label

label_names = ['HS', 'Abusive']

# K-Fold Cross Validation

kfold = StratifiedKFold(n_splits=10, shuffle=True, random_state=42)
kfold

# Logistic Regression

from luwiji.logistic_regression import illustration, demo

illustration.multinomial

demo.loss_curve()

demo.logistic_regression()

demo.loss_plane()

model_logreg = LogisticRegression()

# Randomized Search CV

# Hyperparameter Tuning

hyperparameters_logit =  {
    'penalty' : ['l2', 'l1'],
    'C' :  np.linspace(0, 10, 150),
    'solver' : ['liblinear', 'lbfgs']
}

grid = RandomizedSearchCV(model_logreg, hyperparameters_logit,
                          cv=kfold,
                          n_jobs=-1,
                          verbose=200, return_train_score=True)

# MultiOutput Classifier

clf = MultiOutputClassifier(grid)
clf.fit(X_train_tfidf, y_train)

## fit logistic regression model

# Learning Curve
train_sizes, train_scores, test_scores, fit_times, _ = learning_curve(
    clf,
    X_train_tfidf,
    y_train,
    return_times=True)

data = train_sizes, train_scores, test_scores, fit_times
pd.DataFrame(data)

train_sizes.shape, train_scores.shape, test_scores.shape, fit_times.shape

train_mean = np.mean(train_scores, axis=1)
train_std = np.std(train_scores, axis=1)
test_mean = np.mean(test_scores, axis=1)
test_std = np.std(test_scores, axis=1)

plt.plot(train_sizes, train_mean, label = "Training Score")
plt.plot(train_sizes, test_mean, label = "Cross-Validation Score")

plt.fill_between(train_sizes, train_mean-train_std, train_mean+train_std, color="#DDDDDD")
plt.fill_between(train_sizes, test_mean-test_std, test_mean+test_std, color="#DDDDDD")

plt.title("Learning Curve")
plt.xlabel("Training Size")
plt.ylabel("Accuracy Score")
plt.legend()

clf

# Metrics Classification

y_pred = clf.predict(X_test_tfidf)

## classification report

print(classification_report(y_test, y_pred,target_names=label_names))

## accuracy metrics

print("Accuracy Score Data Training", clf.score(X_train_tfidf, y_train))
print("Accuracy Score Data Testing", clf.score(X_test_tfidf, y_test))

score_lr = accuracy_score(y_pred, y_test)
print(score_lr)

clf.estimators_

from sklearn.neighbors import KNeighborsClassifier

knn = KNeighborsClassifier()

# Hyperparameter Tuning

n_neighbors = range(1, 51, 2)
weights = ['uniform', 'distance']

hyperparameters_knn = dict(n_neighbors=n_neighbors,
                           weights=weights)

grid_knn = RandomizedSearchCV(knn,
                    hyperparameters_knn,
                    cv=kfold,
                    n_jobs=-1,
                    verbose=200)

clf_knn = MultiOutputClassifier(grid_knn).fit(X_train_tfidf, y_train)

y_pred = clf_knn.predict(X_test_tfidf)

print(classification_report(y_test, y_pred,target_names=label_names))

print("Accuracy Score Data Training", clf_knn.score(X_train_tfidf, y_train))
print("Accuracy Score Data Testing", clf_knn.score(X_test_tfidf, y_test))

# Save Model

model_path = "/home/dadandw/scrispy/build_model/skripsi/model/algo_model/model_logreg.jlib"

joblib.dump(clf, model_path)
